Refactor main menu and hand rendering
